+++ cncmark/cncmark/point.py
@@ -79,34 +79,19 @@ def get_shapes(stl_file_path: str, z: float):
     return ground_parallel_lines, ground_parallel_arcs
 
 
-def import_lines(lines: list):
-    cnx = mysql.connector.connect(**MYSQL_CONFIG, database="coord")
-    cursor = cnx.cursor()
-    line_list = [to_line_list(ab) for ab in lines]
+def get_unique_points(lines: list, arcs: list):
+    from .arc import pick_arc_points
 
-    insert_query = "INSERT INTO line (a, b, length) VALUES (%s, %s, %s)"
-
-    try:
-        cursor.executemany(insert_query, line_list)
-    except IntegrityError:
-        print("Error: unable to import lines")
-    cnx.commit()
-    cursor.close()
-    cnx.close()
-
-
-def to_line_list(ab: list):
-    a = ab[0]
-    b = ab[1]
-    length = np.linalg.norm(a - b)
-    return [point_id(a), point_id(b), float(length)]
-
-
-def get_unique_points(lines: list):
     points = []
     for line in lines:
         for point in line:
             points.append(point)
+
+    for arc in arcs:
+        arc_points = pick_arc_points(arc)
+        for point in arc_points:
+            points.append(point)
+
     points = np.array(points)
     unique_points = np.unique(points, axis=0)
     return unique_points
